requirements.txt
streamlit
biopython
matplotlib
stmol
ipywidgets
ipython_genutils
requests
py3Dmol
